1. 更新了目录文件Directory.xml

+++ SimuTraining/util/RecoveryHelper.cs
@@ -15,6 +15,8 @@ namespace SimuTraining.util
 
         public void readLog()
         {
+            if (!File.Exists("app.log"))
+                return;
             String[] allLines = File.ReadAllLines("app.log");
             File.Delete("app.log");
             if (allLines.Length == 0)
